Updated main and right Conkys
Moved the # of active and the total number of processes from the right conky to the left conky and reduced the size of the top 5 headers to make a bit more room for showing all of the network connections (when approaching the max of 18 displayed) on-screen, including up to 6 inbound and 12 outbound connections.  This also meant a realignment of all of the images except for the capslock and numlock indicators.