BIG UPDATE!
This is a MAJOR update to the right conky with the network connections list!  It now occupies the entire vertical part of the right side of the screen instead of resizing, and it will now list up to 18 total connections, including up to 6 inbound and 18 outbound (minus the inbound, of course).  I finally figured out the proper usage of the reverse linefeed character "\" in the conky config, so there are NO MORE VOFFSETS used for fixing layout issues!  Enjoy!