More tweaks & a hardware update!
Replaced my last hard drive with an SSD of approximately double the capacity for my data drive, so the (hwmon) hardware monitoring info changed and thus the main conky.conf required some editing, and I took advantage of the opportunity to change the clock layout and font.  The network, aka right conky, has been further tweaked for better functionality and could now accommodate a 19th connection on a 1080 screen if you don't mind it being beneath the taskbar when it is visible, otherwise, at 18 connections, it is well clear of the taskbar at the bottom, even if it's a fairly big one.