New Year, New Update!
Moved ALL of the round "LED indicators" into the Lua for the system info conky, eliminating the need for the three LED images, as they are now drawn directly via the same lua that draws the bars in the system info conky.  This also makes editing the thresholds where the color goes from green to yellow to red much easier as they are all in a table at the end of the lua.