Efficiency update...
The templates and color definitions had become a mess after over a year of editing, so I went through all 3 conkys and consolidated the color and template definitions for greater efficiency.  This was a pretty significant update for the right conky that monitors the network connections.  The lua bars gained a few segments, too.  It is normal for the system and right conkys to take a few seconds to load.  See the .sh loader script for more info and the suggested loading order if you are having issues.