1/6 Updates!
Added AMD GPU load and frequency info to the CPU conky, and moved the System Info conky down a bit to fit it in!  
I also updated the System Info conky:  I changed the icons for the drives in the storage section and replaced the swap icon. Dispensed with the round LED's for the drives as they are duplicating the effort of the bar graph and text and were therefore quite redundant.  Kept the LED for the swap partition because I like it.