Minor refinements to the MAJOR update!
Ok, so there have been some MAJOR reorganizations of the conkys lately, and this has made some refinements and tweaks necessary to get everything looking and working just right.  This update adds the temperature LED's to the CPU and GPU conky with a more refined temperature threshold for the yellow and red colors.  Added LED's back in for the CPU Socket and Mainboard temps in the System section.  Despite the fact that the system info conky is displaying more data than ever, and in a denser layout, it actually loads 2-3 seconds faster than the one from just 2 versions ago!   I recommend using a startup script, as the system info conky does take approximately 9-10 seconds to load and initialize fully on my system, but once it's running, it performs flawlessly.  There are now well in excess of 130 individual data points being displayed between the 2 left side conkys, thus the long initialization time!