2 Major changes!
Firstly, I changed the background and default colors to grey instead of dark blue for greater universal visibility and to match my desktop themes better, which are mostly gray-based.  Secondly, I have revised the lua for the network conky to use logarithmic scaling for the LED bars so that it is much more reactive to throughput overall and especially at lower speeds!  If you have a decently fast broadband connection, this tweak makes the meters very much more reactive and interesting!!   I also corrected a minor formatting issue and a made a few configuration-efficiency tweaks to the rightconky, which now loads a little faster.  The startup script has been edited for smoother startup on my desktop.