Lua Efficiency Update!
Updated ALL of the lua code to be as efficient as it can be using a couple of different AI models!  This involved more use of lua's caching capabilities and some efficiency tweaks to the code to remove redundancies.  The loader script has been edited for better functionality and to allow time for each conky to load before trying to load the next one.  This script is still a bit of a work in progress.