Refined network detection results...
Refined the code to make the conky look a lot less "broken" when there is no network connection.  Previously, the network icon and the lua bars would not behave properly when there was no connection.  Now, the network icon stays put and gets the red light, but the text to the right is replaced with a red flashing "NO CONNECTION!" and the lua upload and download bars will not draw at all when there is no connection now, which is good because all of the surrounding text is also not there when there is no connection, making it look much better and less "broken".