Visual updates!
Updated the system info conky and lua to create the darkened background behind every other entry in the top 10 processes using RAM and CPU at the bottom of the config.  Also edited the Network conky to achieve the same effect for every other network connection in the list of up to 25 connections.  The desired effect mimics the effect used to aid in reading spreadsheets where every other line is shaded some color.