Visual update!
Added a blue semi-transparent background bar behind every other connection in the network conky, every other process in the RAM and CPU lists, and did a visual update to the CPU conky, adding a bit-o-ANSI and CPU Core #'s along with some background shading for better readability.  The goal was to emulate the effect used on old-school paper spreadsheets whereby every other line is shaded, making it easier to read  properly.  This shading is actually done via the lua files and are called in a pre-draw loop from within the conky so the shading does not affect the text colors.  Enjoy!