MAJOR NEW Connection List Update!
I have added connection condensation to the network connection list!  What this means is that if you have 4 https connections to 192.168.1.1, then instead if using 4 slots in the list, it should now only use 1 slot, and add " (4)" after the IP address to indicate that there are 4 active connections of that type with that ip.  This keeps applications like Steam, which opens well over a dozen connections, from utterly swamping the network connection list.