KWin-Zones Script Settings!
This text file contains my settings for the KDE Plasma Desktop Environment KWin Script called "Zones", which allows for the definition of the areas for windows to snap/tile to.  With the proper Kwin settings, Zones allows you to select from virtually any number of layout combinations (that you define using settings such as those I have provided in this file) when moving/resizing windows!
In this case, these settings mean that by default, all windows open to cover the center between the conkys that line either side of my screen.  When using KDE Plasma @ 1080p resolution, windows will not normally open to cover the conkys, but will fit neatly between them except for the Dual HD layout, which is for putting 2 HD videos on the screen side by side at the bottom, which only partially-covers the conkys.